feat: button.html to button.tsx and render by  `renderToString` in `react-dom`
