+++ .storybook/main.ts
@@ -1,7 +1,7 @@
 import type { StorybookConfig } from "@storybook/html-vite";
 
 const config: StorybookConfig = {
-  stories: ["../src/**/*.mdx", "../src/**/*.stories.ts"],
+  stories: ["../src/**/*.mdx", "../src/**/*.stories.tsx"],
   addons: ["@storybook/addon-essentials", "@storybook/addon-interactions"],
   framework: {
     name: "@storybook/html-vite",
