+++ package.json
@@ -25,6 +25,9 @@
   "dependencies": {
     "@tailwindcss/cli": "4.0.0-beta.9",
     "@tailwindcss/vite": "4.0.0-beta.9",
+    "clsx": "^2.1.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "tailwindcss": "4.0.0-beta.9",
     "vite": "^6.0.7"
   },
@@ -34,11 +37,13 @@
     "@storybook/addon-essentials": "^8.4.7",
     "@storybook/addon-interactions": "^8.4.7",
     "@storybook/blocks": "^8.4.7",
-    "@storybook/html": "^8.4.7",
     "@storybook/html-vite": "^8.4.7",
     "@storybook/test": "^8.4.7",
     "@tailwindcss/postcss": "0.0.0-development.1",
     "@tsconfig/strictest": "^2.0.5",
+    "@tsconfig/vite-react": "^3.4.0",
+    "@types/react": "^19.0.7",
+    "@types/react-dom": "^19.0.3",
     "chromatic": "^11.22.2",
     "prettier": "^3.4.2",
     "prettier-plugin-tailwindcss": "^0.6.9",
